move common data_classes to another file
